+++ lib/src/model/device/virtual/type.dart
