+++ lib/src/model/device/virtual/virtual_device.dart
@@ -3,7 +3,7 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:switchbot_api_dio/src/model/converter.dart';
 import 'package:switchbot_api_dio/src/model/device/device_collection.dart';
 
-export 'virtual_device_type.dart';
+export 'type.dart';
 
 part 'virtual_device.freezed.dart';
