+++ lib/src/model/model.dart
@@ -1,3 +1,5 @@
+export 'common/enum.dart';
+export 'common/light_color.dart';
 export 'control/device_command.dart';
 export 'device/device_collection.dart';
 export 'scene/manual_scene.dart';
