+++ lib/src/model/status/cleaner.dart
@@ -75,24 +75,3 @@ class _DeviceStatusVacuumCleaner
   factory _DeviceStatusVacuumCleaner.fromJson(Map<String, dynamic> json) =>
       _$_DeviceStatusVacuumCleanerFromJson(json);
 }
-
-/// Status of online/offline
-enum OnlineStatus { online, offline }
-
-/// Status of 'Robot Vacuum Cleaner S1 **' device
-enum VacuumCleanerStatus {
-  standBy('StandBy'),
-  clearing('Clearing'),
-  paused('Paused'),
-  gotoChargeBase('GotoChargeBase'),
-  charging('Charging'),
-  dormant('Dormant'),
-  inTrouble('InTrouble'),
-  inRemoteControl('InRemoteControl'),
-  inDustCollecting('InDustCollecting'),
-  ;
-
-  const VacuumCleanerStatus(this.value);
-
-  final String value;
-}
