+++ lib/src/model/status/device_status.dart
@@ -1,5 +1,7 @@
 // ignore_for_file: invalid_annotation_target
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:switchbot_api_dio/src/model/common/enum.dart';
+import 'package:switchbot_api_dio/src/model/common/light_color.dart';
 import 'package:switchbot_api_dio/src/model/converter.dart';
 import 'package:switchbot_api_dio/src/model/device/physical/type.dart';
